download checks

+++ prep/download.py
@@ -4,16 +4,17 @@ from os.path import join as osj
 import pandas as pd
 import pdb
 
-hosting_address = 'https://www.robots.ox.ac.uk/~vgg/research/condensed-movies/data/'
+hosting_address = 'https://www.robots.ox.ac.uk/~vgg/research/condensed-movies/data'
 
 
 def download_features(data_dir):
-    cmd = 'wget {}/features.zip -P {}; unzip {}/features.zip'.format(hosting_address, data_dir, data_dir)
+    cmd = 'wget {}/features.zip -P {}; unzip {}/features.zip -d {}'.format(hosting_address, data_dir, data_dir, data_dir)
     os.system(cmd)
 
 
 def download_facetracks(data_dir):
-    cmd = 'wget {}/facetracks.zip -P {}; unzip {}/facetracks.zip'.format(hosting_address, data_dir, data_dir)
+    cmd = 'wget {}/facetracks.zip -P {}/facetracks.zip; unzip {}/facetracks.zip -d {}'.format(hosting_address, data_dir,
+                                                                                              data_dir, data_dir)
     os.system(cmd)
 
 
@@ -32,21 +33,18 @@ def youtube_download(data_dir):
         id_fp = osj(id_dir, file)
         cmd = 'youtube-dl --config-location youtube-dl.conf -o "{}" -a "{}"'.format(output_fmt, id_fp)
         os.system(cmd)
-        break
 
     # trim advertisement outro from video.
     trim = None
     while trim not in ['y', 'n']:
         trim = str(input(
-            "\n #### Do you want to trim the videos (y/n) ?###\n\
-            This removes the advertisements (unrelated to the film), and only needs to be done once per download. "))
+            "\nDo you want to trim the videos (y/n)?\nThis removes the advertisements (unrelated to the film), and only needs to be done once per download."))
 
         if trim not in ['y', 'n']:
             print('Please type "y" or "n"')
 
     if trim == "y":
         trim_video_outro(video_dir)
-
 
     # check for failed downloads: this can be due to...
     # i) geographical restrictions
@@ -60,17 +58,18 @@ def trim_video_outro(video_dir, video_ext='.mkv'):
     tmp_fp = osj(video_dir, 'tmp' + video_ext)
     for root, subdir, files in os.walk(video_dir):
         for file in files:
-            if file.endswith(video_ext):
+            if file.endswith(video_ext) and file != 'tmp' + video_ext:
                 videoid = file.split(video_ext)[0]
+                pdb.set_trace()
                 if videoid not in duration_data.index:
                     raise ValueError("Videoid not found, video files should be in format {VIDEOID}.mkv")
 
                 video_fp = osj(root, file)
-                new_duration = duration_data['videoid']
+                new_duration = duration_data.loc[videoid]['duration']
 
                 # create tmp for untrimmed
                 os.system('cp {} {}'.format(video_fp, tmp_fp))
-                cmd = ' ffmpeg -y -ss 0 -i {} -t {} -c copy {}'.format(tmp_fp, new_duration, video_fp)
+                cmd = 'ffmpeg -y -ss 0 -i {} -t {} -c copy {}'.format(tmp_fp, new_duration, video_fp)
                 os.system(cmd)
                 os.remove(tmp_fp)
 
@@ -78,21 +77,21 @@ def trim_video_outro(video_dir, video_ext='.mkv'):
 def check_missing_vids(video_dir, video_ext='.mkv'):
     missing_ids = []
     clips_data = pd.read_csv('../metadata/clips.csv').set_index('videoid')
-    for idx, row in clips_data:
-        videoid = row.index
+    for idx, row in clips_data.iterrows():
+        videoid = row.name
         upload_year = row['upload_year']
-        video_fp = osj(video_dir, upload_year, videoid + video_ext)
+        video_fp = osj(video_dir, str(int(upload_year)), videoid + video_ext)
         if not os.path.isfile(video_fp):
             missing_ids.append(videoid)
 
-    success = len(missing_ids) * 100 / len(clips_data)
-    print('%.2f %% of clips downloaded successfully' % success)
+    success = (1 - len(missing_ids) / len(clips_data)) * 100
+    print('=======================================================\n%.2f %% of clips downloaded successfully' % success)
     if success == 100:
         pass
     elif success < 100:
         print(
-            '%d clips failed to download. This is likely due to geographical restrictions.\n\
-            Contact maxbain@robots.ox.ac.uk if this is an issue.' % len(missing_ids))
+            '%d clips failed to download. This is likely due to geographical restrictions.\n-->Contact maxbain@robots.ox.ac.uk if this is an issue.' % len(
+                missing_ids))
 
     with open('missing_videos.out', 'w') as fid:
         for mid in missing_ids:
